Move common GTK Backend code into a separate file.
